refactor: rename ThemalijnenBlock to ThemelinesBlock #142
